Pulling out repo manager
